+++ app/routes/devices.$ip_addr.tsx
@@ -1,0 +1,12 @@
+import { Outlet, Link, useParams } from "@remix-run/react";
+
+export const handle = {
+  Breadcrumb: () => {
+    const { ip_addr } = useParams();
+    return <Link to={`/devices/${ip_addr}/hardware`}>{ip_addr}</Link>;
+  },
+};
+
+export default function SomeParent() {
+  return <Outlet />;
+}
